Use -lcrypto.
I'm not happy with this change at all, but it does allow the build to work.

I want to make this an optional thing, default to on. If it is off
pe.imphash() should return UNDEFINED.

I also want to scope this down a bit so yarac doesn't end up getting
-lcrypto too.

I imagine this will need to be revisited by someone with a better
understanding of the build.