Properly calculate the next certificate entry.
There is a lot of confusion around this. The specification says it must be
aligned to an 8 byte boundary and the length field does not include the
structure itself. Most binaries adhere to this but I've seen some for which
the length includes the structure and the padding.

There needs to be more investigation on the proper way to handle the (rare)
case of a file with multiple signatures.