Include "offset" and "length" in pe.rich_signature and remove "start"
By including "offset" and "length" we can compute a hash for the rich signature, which can be a useful way of detecting files. Example:

 hash.md5(pe.rich_signature.offset, pe.rich_signature.length) == "c5d73827a4a2b4f8ca1e6af5de76a7dd"

"start" is not that useful, because it "key" xored with a constant. If you need the value for "start" you can do uint32be(pe.rich_signature.offset)