Use the OptionalHeader.Magic value.
When determining if a file is PE32 or PE32+ you need to use the Magic
value in the OptionalHeader, not Fileheader.Machine. If you incorrectly
use Fileheader.Machine then you will improperly parse 64bit files for
non-AMD64 architectures. In particular I have seen this problem when
looking at an ARM64 binary. The resource directories were all off by
some amount because it was using the wrong OptionalHeader because of the
wrong check.